Fix type of Instrument.UseDuration

+++ Sharpmine.Domain/Registries/Dynamic/Entries/Instrument.cs
@@ -4,6 +4,6 @@ namespace Sharpmine.Domain.Registries.Dynamic.Entries;
 
 public record Instrument(
     Identifier SoundEvent,
-    int UseDuration,
+    float UseDuration,
     float Range,
     TextComponent Description);
